Refactor the importer classes to move things into the base class instead of the command.
