+++ polling_stations/apps/data_collection/management/commands/__init__.py
@@ -1,17 +1,55 @@
-import os
+"""
+Defines the base importer classes to override
+"""
+import json
 import glob
+import os
+import shapefile
 
 from django.core.management.base import BaseCommand
 from django.contrib.gis import geos
+from django.contrib.gis.geos import Point, GEOSGeometry
 
 from councils.models import Council
+from pollingstations.models import PollingStation, PollingDistrict
 
 class BaseImporter(BaseCommand):
     pass
 
 class BaseShpImporter(BaseCommand):
-    stations_name = "polling_places"
+    srid = 27700
+
+    council_id     = None
+    stations_name  = "polling_places"
     districts_name = "polling_districts"
+
+    def import_data(self):
+        self.import_polling_districts()
+        self.import_polling_stations()
+
+
+    def import_polling_districts(self):
+        sf = shapefile.Reader("{0}/{1}".format(
+            self.base_folder_path,
+            self.districts_name
+            ))
+        for district in sf.shapeRecords():
+            district_info = self.district_record_to_dict(district.record)
+            # import ipdb; ipdb.set_trace()
+
+            geojson = json.dumps(district.shape.__geo_interface__)
+            poly = self.clean_poly(GEOSGeometry(geojson, srid=self.srid))
+
+            district_info['area'] = poly
+            # import ipdb; ipdb.set_trace()
+            self.add_polling_district(district_info)
+
+    def add_polling_district(self, district_info):
+        PollingDistrict.objects.update_or_create(
+            council=self.council,
+            internal_council_id=district_info.get('internal_council_id', None),
+            defaults=district_info,
+        )
 
     def clean_poly(self, poly):
         # print wkt
@@ -42,9 +80,31 @@ class BaseShpImporter(BaseCommand):
         # print WKT
         return poly
 
+    def add_polling_station(self, station_info):
+        PollingStation.objects.update_or_create(
+            council=self.council,
+            internal_council_id=station_info['internal_council_id'],
+            defaults=station_info,
+        )
+
+    def import_polling_stations(self):
+        sf = shapefile.Reader("{0}/{1}".format(
+            self.base_folder_path,
+            self.stations_name
+            ))
+        for station in sf.shapeRecords():
+            station_info = self.station_record_to_dict(station.record)
+            station_info['location'] = Point(
+                *station.shape.points[0],
+                srid=self.srid)
+            self.add_polling_station(station_info)
+
+
 
     def handle(self, *args, **kwargs):
-        self.council_id = args[0]
+        if self.council_id is None:
+            self.council_id = args[0]
+            
         self.council = Council.objects.get(pk=self.council_id)
         self.base_folder_path = os.path.abspath(
          glob.glob('data/{0}-*'.format(self.council_id))[0]
@@ -52,4 +112,9 @@ class BaseShpImporter(BaseCommand):
 
 
         self.import_data()
+
+
+
+
+
